skeleton of insert
turns out generic_array won't play nice with
quickcheck, so out it goes. also, clone is
a bit of a pain with arrays, so I've hand
coded it up for a specific array length.
Later on I can write a macro for
generalizing this, but for now, this is good
enough.